+++ docs/changelog.md
@@ -89,6 +89,10 @@
 
 ## 更新说明详情
 
+1.18.35 fix: fix stock_ipo_ths interface
+
+	1. 修复 stock_ipo_ths 接口
+
 1.18.34 fix: fix fund_fh_em interface
 
 	1. 修复 fund_fh_em 接口
@@ -6169,6 +6173,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.35 fix: fix stock_ipo_ths interface
 
 1.18.34 fix: fix fund_fh_em interface